Algae shoulder safe angle and other logic fixes (#60)
* Added a reef safe angle for algae.
Also renamed the current reef angle to be for coral.
Made the SuperStructure goToPosition and followPosition both take in a Supplier<Gamepiece> so it knows which limit to use; the Coral or Algae.
Made all there usages pass in a Supplier<Gamepiece>.
A couple formatting things.

* Fixes and improvements
Made the constants better for the positions.
Fixed the go-to default positions when we have an algae.
Made the follow positions not require gamepiece because I can just use the haveAlgae that we already have.
Made the wrist wait to move until the shoulder is retracted so that it doesn't push it downwards.
Made the elevator logic work so that it properly waits for the shoulder when we have an algae.
Removed unnecessary Constants and variables.

* Constants naming fix: SHOULDER_MAX_MOVING_ANGLE -> SHOULDER_MAX_MOVING_OFFSET