Add coral station position with coral in the way (#75)
* Logic for lining up with coral station with Coral in the way.
Added a Robot Mode for it.
And also corresponding Constants for the positions.
Pass in which one you are doing into the align with Coral Station method, if you don't pass in one of the Coral Station robotModes it just assumes the normal Coral Station.

* deleting unnecessities

got rid of unnecessities

* fix issue

got rid of a null

* Coral station with coral in the way

all working code minus the actual values for it

* got the values

got the values for the coral in the way and added it to smart dashboard

* merge issues and brake mode

* Made the robotModeCS to a boolean

The robotModeCS is only ever two values so converted it to a boolean.
Changed the SmartDashBoard boolean to be a single value since it is a boolean.

---------

Co-authored-by: Kaitlin Moody <kaitlin@mymoodyfamily.com>